refactor: data to renderData
