+++ libs/frontend/modules/builder/src/renderer/pipes/test/extraElementPropsPipe.spec.ts
@@ -1,6 +1,6 @@
 import { extraElementPropsPipe } from '../extraElementPropsPipe'
 import { RenderContext } from '../types'
-import { elementToRender } from './data'
+import { elementToRender } from './renderData'
 import { EndPipeOutput } from './types'
 import { endPipe } from './utils'
