+++ libs/frontend/modules/builder/src/renderer/pipes/test/index.ts
@@ -1,2 +1,2 @@
-export * from './data'
+export * from './renderData'
 export * from './types'
