+++ libs/frontend/modules/builder/src/renderer/pipes/test/propModifiersPipeline.spec.ts
@@ -4,7 +4,7 @@ import { ReactElement } from 'react'
 import { RenderProps } from '../../../store'
 import { propModifiersPipeline } from '../renderPipeline'
 import { RenderContext } from '../types'
-import { elementToRender } from './data'
+import { elementToRender } from './renderData'
 import { EndPipeOutput } from './types'
 import { endPipe } from './utils'
