+++ libs/frontend/modules/builder/src/renderer/pipes/test/propsPipeline.spec.ts
@@ -1,6 +1,6 @@
 import { propsPipeline } from '../renderPipeline'
 import { RenderContext } from '../types'
-import { elementToRender } from './data'
+import { elementToRender } from './renderData'
 import { EndPipeOutput } from './types'
 import { endPipe } from './utils'
